Moved WAV lib to separate folder 'lib'
